hackaton/backend/requirements.txt
openai
fastapi
uvicorn
python-dotenv
python-multipart
pillow